Projekt końcowy/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ale-py==0.8.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.2
Automat==20.2.0
AutoROM.accept-rom-license==0.6.1
AutoROM==0.4.2
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.11.2
bertopic==0.15.0
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
confection==0.0.4
constantly==15.1.0
contourpy==1.0.7
cryptography==38.0.1
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
emoji==2.4.0
en-core-web-md==3.5.0
en-core-web-sm==3.5.0
executing==1.2.0
fastapi==0.87.0
fastjsonschema==2.16.3
feedparser==6.0.10
filelock==3.8.0
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.5
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gym-notices==0.0.8
gym==0.26.2
h11==0.14.0
hdbscan==0.8.29
html5lib==1.1
httptools==0.5.0
huggingface-hub==0.14.1
hyperlink==21.0.0
idna==3.4
imageio-ffmpeg==0.4.8
imageio==2.31.0
importlib-resources==5.12.0
incremental==22.10.0
inflect==6.0.2
ipykernel==6.21.2
ipython-genutils==0.2.0
ipython==8.11.0
ipywidgets==8.0.4
isoduration==20.11.0
itemadapter==0.7.0
itemloaders==1.0.6
jaraco.collections==4.0.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-console==6.6.2
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server-terminals==0.4.4
jupyter-server==2.3.0
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterthemes==0.20.0
kaleido==0.2.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy-loader==0.2
lesscpy==0.15.1
llvmlite==0.40.0
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.7.0
mistune==2.0.5
mlxtend==0.22.0
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
multitasking==0.0.11
murmurhash==1.0.9
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook-shim==0.2.2
notebook==6.5.2
numba==0.57.0
numpy==1.23.5
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pl-core-news-md==3.5.0
pl-core-news-sm==3.5.0
platformdirs==3.0.0
plotly==5.14.1
ply==3.11
pmdarima==2.0.3
portend==3.1.0
praw==7.7.0
prawcore==2.3.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.6
pygame==2.1.0
Pygments==2.14.0
pynndescent==0.5.10
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
regex==2023.3.23
requests-file==1.5.1
requests==2.28.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.7.0
seaborn==0.12.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==21.1.0
setuptools==58.1.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy==3.5.1
srsly==2.4.6
stack-data==0.6.2
starlette==0.21.0
statsmodels==0.13.5
sympy==1.12
tempora==5.2.1
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.13.3
torch-directml==0.2.0.dev230426
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.29.1
twisted-iocpsupport==1.0.2
Twisted==22.8.0
typer==0.7.0
typing-extensions==4.4.0
typing-inspect==0.9.0
umap-learn==0.5.3
update-checker==0.18.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.19.0
w3lib==2.0.1
wasabi==1.1.1
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
xxhash==3.2.0
yarl==1.8.2
yfinance==0.2.12
zc.lockfile==3.0.post1
zope.interface==5.5.0
